pulled out voting intent functionality and put it within its own utils class
