+++ mycity/mycity/intents/voting_intent.py
@@ -7,6 +7,7 @@ import mycity.intents.speech_constants.voting_intent as speech_constants
 from streetaddress import StreetAddressParser
 from mycity.mycity_response_data_model import MyCityResponseDataModel
 import mycity.utilities.arcgis_utils as arcgis_utils
+import mycity.utilities.voting_utils as vote_utils
 import requests
 import logging
 
@@ -31,8 +32,8 @@ def get_polling_location(mycity_request):
     current_address += ", Boston, MA"
     candidates = arcgis_utils.geocode_address_candidates(current_address)
     top_candidate = arcgis_utils.select_top_address_candidate(candidates)
-    ward_precinct = arcgis_utils.get_ward_precinct_info(top_candidates)
-    poll_location = arcgis_utils.get_polling_location(ward_precinct)
+    ward_precinct = vote_utils.get_ward_precinct_info(top_candidates)
+    poll_location = vote_utils.get_polling_location(ward_precinct)
     output_speech = speech_constants.LOCATION_SPEECH.format(poll_location["Location Name"], poll_location["Location Address"])
     mycity_response.output_speech = output_speech
     mycity_response.reprompt_text = None
